REFACTOR: 파일 전체 구조
