+++ src/main/java/bsm/devcoop/oring/domain/agenda/AgendaId.java
@@ -1,4 +1,4 @@
-package bsm.devcoop.oring.domain.vote;
+package bsm.devcoop.oring.domain.agenda;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -12,6 +12,6 @@ import java.time.LocalDate;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class AgendaId implements Serializable {
-    private int agendaId;
-    private LocalDate conferenceId;
+  private int agendaId;
+  private LocalDate conferenceId;
 }
