+++ src/main/java/bsm/devcoop/oring/domain/conference/repository/ConferenceRepository.java
@@ -1,7 +1,6 @@
-package bsm.devcoop.oring.domain.vote.repository;
+package bsm.devcoop.oring.domain.conference.repository;
 
-import bsm.devcoop.oring.domain.vote.Conference;
-import bsm.devcoop.oring.domain.vote.Vote;
+import bsm.devcoop.oring.domain.conference.Conference;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
