+++ src/main/java/bsm/devcoop/oring/domain/vote/VoteId.java
@@ -1,6 +1,6 @@
 package bsm.devcoop.oring.domain.vote;
 
-
+import bsm.devcoop.oring.domain.agenda.AgendaId;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -8,12 +8,10 @@ import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
 
-import java.time.LocalDate;
-
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Embeddable
 public class VoteId implements Serializable {
-    private AgendaId agendaId;
-    private char studentId;
+  private AgendaId agendaId;
+  private char studentId;
 }
