+++ src/main/java/bsm/devcoop/oring/domain/vote/repository/VoteRepository.java
@@ -6,5 +6,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface VoteRepository extends JpaRepository<Vote, String> {
+
 }
