+++ src/main/java/bsm/devcoop/oring/domain/vote/service/VoteService.java
@@ -1,10 +1,8 @@
 package bsm.devcoop.oring.domain.vote.service;
 
-import bsm.devcoop.oring.domain.vote.Student;
-import bsm.devcoop.oring.domain.vote.Vote;
+import bsm.devcoop.oring.domain.user.User;
 import bsm.devcoop.oring.domain.vote.presentation.dto.VotingRequestDto;
-import bsm.devcoop.oring.domain.vote.repository.StudentRepository;
-import bsm.devcoop.oring.domain.vote.repository.VoteRepository;
+import bsm.devcoop.oring.domain.user.presentation.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -15,12 +13,12 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Slf4j
 public class VoteService {
-    private final StudentRepository studentRepository;
+  private final UserRepository userRepository;
 
-    // 투표하기 ( 현재 학번만 return )
-    @Transactional
-    public char vote(VotingRequestDto requestDto) {
-        Student student = studentRepository.findByStuNumber(requestDto.getStuNumber());
-        return student.getStuNumber();
-    }
+  // 투표하기 ( 현재 학번만 return )
+  @Transactional
+  public char vote(VotingRequestDto requestDto) {
+    User user = userRepository.findByStuNumber(requestDto.getStuNumber());
+    return user.getStuNumber();
+  }
 }
