+++ src/main/java/bsm/devcoop/oring/domain/vote/types/VoteCode.java
@@ -1,6 +1,6 @@
 package bsm.devcoop.oring.domain.vote.types;
 
 public enum VoteCode {
-    Y,
-    N
+  Y,
+  N
 }
