+++ src/main/resources/application.yml
@@ -4,8 +4,8 @@ spring:
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
     url: jdbc:mysql://localhost:3306/test?serverTimezone=UTC&characterEncoding=UTF-8
-    password: 1234
-    username: root
+    password: ${DB_PASSWORD}
+    username: ${DB_USERNAME}
   jpa:
     hibernate:
       ddl-auto: create
